preliminary version of ELinkGetResponseModel made, using ELinkAPI  RecordResponse
