+++ mpcite_legacy/__init__.py
@@ -1,0 +1,8 @@
+"""The Materials Project's Citation Framework"""
+
+__version__ = "1.0.0"
+__url__ = "https://github.com/materialsproject/MPCite"
+__author__ = "Patrick Huck & Michael Wu"
+__email__ = "phuck@lbl.gov"
+__copyright__ = "Copyright 2020, The Materials Project"
+__maintainer__ = "Patrick Huck"
