+++ mpcite_legacy/config_ipynb.txt
